+++ protected/humhub/components/bootstrap/SettingsLoader.php
@@ -19,6 +19,7 @@ class SettingsLoader implements BootstrapInterface
         $this->setMailerConfig($app);
         $this->setUserConfig($app);
         $this->setCacheConfig($app);
+        $this->setParams($app);
     }
 
     private function updateComponentDefinition($app, $component, $definition)
@@ -136,5 +137,12 @@ class SettingsLoader implements BootstrapInterface
                 'keyPrefix' => $app->id,
             ]));
         }
+    }
+
+    protected function setParams($app)
+    {
+        $app->name = $app->settings->get('name');
+        $app->params['installed'] = $app->settings->get('installed');
+        $app->params['horImageScrollOnMobile'] = $app->settings->get('horImageScrollOnMobile');
     }
 }